_rayso_api/requirements.txt
selenium
typer
pillow
rich
requests
beautifulsoup4